dom/tests/mochitest/bugs/test_bug463000.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=463000
-->
<head>
  <title>Test for Bug 463000</title>
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8" />
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=463000">Mozilla Bug 463000</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<iframe src="http://example.org/tests/dom/tests/mochitest/bugs/iframe_bug463000.html"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 463000 **/

SimpleTest.waitForExplicitFinish();

function receiveMessage(evt)
{
  is(evt.data, "child-response", "got wrong response");

  SimpleTest.finish();
}

window.addEventListener("message", receiveMessage, false);

</script>
</pre>
</body>
</html>
